rename: 예약 확인/완료 모달창 이름 수정 및 폴더 이동 (#62)
