+++ src/app/camp/detail/[id]/_styles/ConfirmModal.module.css
@@ -5,8 +5,9 @@
   position: absolute;
   top: 40%;
   left: 30%;
+  border: 1px solid #999999;
   border-radius: 1rem;
-  background-color: #f7f8f8;
+  background-color: #ffffff;
   text-align: center;
   z-index: 999;
 }
